requirements.txt
Django==2.1.2
django-heroku==0.3.1
djangorestframework==3.8.2
gunicorn==19.9.0
psycopg2==2.7.5
requests==2.19.1
